requirements.txt
aiobotocore==1.4.2
aiohttp==3.8.4
aiohttp-apispec==2.2.1
aiohttp-cors==0.7.0
aiohttp-session
aioitertools==0.11.0
aioredis
aiosignal==1.3.1
alembic==1.11.1
apispec==3.3.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
botocore==1.20.106
bs64==0.0.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==35.0.0
fernet==1.0.1
frozenlist==1.3.3
greenlet==2.0.2
hiredis==2.2.3
idna==3.4
iniconfig==2.0.0
Jinja2==2.11.3
jmespath==0.10.0
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.14.1
marshmallow-dataclass==8.5.3
multidict==5.2.0
mypy-extensions==1.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
py==1.11.0
pyaes==1.6.1
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pytest==6.2.5
pytest-aiohttp==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
async-sqlalchemy
toml==0.10.2
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
webargs==5.5.3
wrapt==1.15.0
yarl==1.9.2
pydantic[email]
Pillow